Handle event ids greater than int32.maxvalue (#1940)
Fixes https://github.com/octokit/octokit.net/issues/1939

+++ Octokit/Models/Response/EventInfo.cs
@@ -11,7 +11,7 @@ namespace Octokit
     {
         public EventInfo() { }
 
-        public EventInfo(int id, string nodeId, string url, User actor, User assignee, Label label, EventInfoState @event, string commitId, DateTimeOffset createdAt)
+        public EventInfo(long id, string nodeId, string url, User actor, User assignee, Label label, EventInfoState @event, string commitId, DateTimeOffset createdAt)
         {
             Id = id;
             NodeId = nodeId;
@@ -27,7 +27,7 @@ namespace Octokit
         /// <summary>
         /// The id of the issue/pull request event.
         /// </summary>
-        public int Id { get; protected set; }
+        public long Id { get; protected set; }
 
         /// <summary>
         /// GraphQL Node Id
